backend/eshop/requirements/production.txt
Django==3.0.8
djangorestframework==3.10.3
